requirements.txt
#dependencies to be install
astapi
uvicorn
pandas
joblib
pydantic
gradio
requests
